requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pillow>=10.0.0
numpy>=1.24.0
supabase==1.3.0
python-dotenv==1.0.0
requests==2.31.0
face-recognition==1.3.0
dlib-binary==19.24.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pillow>=10.0.0
numpy>=1.24.0
supabase==1.3.0
python-dotenv==1.0.0
opencv-python-headless>=4.8.0
insightface==0.7.3
onnxruntime-cpu>=1.16.0
requests==2.31.0
psutil>=5.9.0